minor fixes and clean up rendering config for jupyter and add function to set global jupyter renderer for section and geometries.
