+++ examples/experiments/rendering_pygfx/render_boxes.py
@@ -1,7 +1,8 @@
 import pathlib
+
 import trimesh
 
-from ada.visit.render_backend import SqLiteBackend, MeshInfo
+from ada.visit.render_backend import SqLiteBackend
 from ada.visit.render_pygfx import RendererPyGFX
 
 glb_file = pathlib.Path(__file__).parent.parent.parent.parent / "files/gltf_files" / "boxes_merged.glb"
