+++ examples/experiments/rendering_pygfx/websockets_send_to_process.py
@@ -1,5 +1,6 @@
 import io
 import time
+
 from websockets.sync.client import connect
 
 import ada
