+++ examples/experiments/rendering_pygfx/websockets_start_process.py
@@ -1,4 +1,3 @@
-from ada.config import logger
 from ada.visit.render_pygfx import standalone_viewer
 
 # logger.setLevel("INFO")
