+++ src/ada/__init__.py
@@ -8,6 +8,7 @@ from ada import fem
 from ada.api.beams import Beam, BeamRevolve, BeamSweep, BeamTapered
 from ada.api.curves import ArcSegment, CurvePoly2d, CurveRevolve, LineSegment
 from ada.api.fasteners import Bolts, Weld
+from ada.api.groups import Group
 from ada.api.nodes import Node
 from ada.api.piping import Pipe, PipeSegElbow, PipeSegStraight
 from ada.api.plates import Plate
@@ -22,8 +23,7 @@ from ada.api.primitives import (
     PrimSweep,
     Shape,
 )
-from ada.api.spatial import Part, Assembly
-from ada.api.groups import Group
+from ada.api.spatial import Assembly, Part
 from ada.api.transforms import Instance, Placement, Transform
 from ada.api.user import User
 from ada.api.walls import Wall
@@ -34,6 +34,7 @@ from ada.geom.placement import Direction
 from ada.geom.points import Point
 from ada.materials import Material
 from ada.sections import Section
+from ada.visit.config import set_jupyter_part_renderer
 
 if TYPE_CHECKING:
     import ifcopenshell
@@ -149,4 +150,5 @@ __all__ = [
     "Weld",
     "Units",
     "fem",
+    "set_jupyter_part_renderer",
 ]
