+++ src/ada/api/spatial/__init__.py
@@ -1,4 +1,4 @@
-from .part import Part
 from .assembly import Assembly
+from .part import Part
 
 __all__ = ["Part", "Assembly"]
