+++ src/ada/cadit/ifc/write/write_beams.py
@@ -16,11 +16,11 @@ from ada.cadit.ifc.utils import (
     ifc_p,
     to_real,
 )
-from ada.core.utils import create_guid
 from ada.cadit.ifc.write.geom.solids import extruded_area_solid
 from ada.cadit.ifc.write.write_curves import write_curve_poly
 from ada.config import Settings
 from ada.core.constants import O
+from ada.core.utils import create_guid
 
 if TYPE_CHECKING:
     from ifcopenshell import file as ifile
