+++ src/ada/cadit/ifc/write/write_ifc.py
@@ -8,11 +8,7 @@ import ifcopenshell.geom
 
 from ada.base.changes import ChangeAction
 from ada.cadit.ifc.read.reader_utils import get_ifc_body
-from ada.cadit.ifc.utils import (
-    add_negative_extrusion,
-    write_elem_property_sets,
-)
-from ada.core.utils import create_guid
+from ada.cadit.ifc.utils import add_negative_extrusion, write_elem_property_sets
 from ada.cadit.ifc.write.write_beams import update_ifc_beam, write_ifc_beam
 from ada.cadit.ifc.write.write_fasteners import write_ifc_fastener
 from ada.cadit.ifc.write.write_instances import write_mapped_instance
@@ -28,6 +24,7 @@ from ada.cadit.ifc.write.write_spatial_elements import (
 )
 from ada.cadit.ifc.write.write_wall import write_ifc_wall
 from ada.config import logger
+from ada.core.utils import create_guid
 
 if TYPE_CHECKING:
     from ada import Beam, Boolean, Material, Part, Pipe, Plate, Section, Shape, Wall
