+++ src/ada/cadit/ifc/write/write_pipe.py
@@ -13,9 +13,9 @@ from ada.cadit.ifc.utils import (
     to_real,
     write_elem_property_sets,
 )
-from ada.core.utils import create_guid
 from ada.config import logger
 from ada.core.constants import O, X, Z
+from ada.core.utils import create_guid
 from ada.core.vector_transforms import global_2_local_nodes
 from ada.core.vector_utils import (
     angle_between,
