+++ src/ada/cadit/ifc/write/write_plates.py
@@ -1,7 +1,5 @@
 from ada import Plate
-from ada.cadit.ifc.utils import (
-    add_colour,
-)
+from ada.cadit.ifc.utils import add_colour
 from ada.cadit.ifc.write.geom.placement import ifc_placement_from_axis3d
 from ada.cadit.ifc.write.geom.solids import extruded_area_solid
