+++ src/ada/cadit/ifc/write/write_spatial_elements.py
@@ -3,10 +3,7 @@ from __future__ import annotations
 from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
-from ada.cadit.ifc.utils import (
-    create_local_placement,
-    write_elem_property_sets,
-)
+from ada.cadit.ifc.utils import create_local_placement, write_elem_property_sets
 from ada.core.utils import create_guid
 from ada.fem.formats.ifc.writer import to_ifc_fem
