+++ src/ada/fem/formats/abaqus/read/read_orientations.py
@@ -1,4 +1,4 @@
-from ada.fem import Csys, FEM
+from ada.fem import FEM, Csys
 
 from . import cards
 from .helper_utils import get_set_from_assembly
