+++ src/ada/fem/formats/code_aster/write/write_materials.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from itertools import chain
 from typing import TYPE_CHECKING
 
-from ada.materials.concept import Material
 from ada.config import logger
+from ada.materials.concept import Material
 
 if TYPE_CHECKING:
     from ada.api.spatial import Assembly
