+++ src/ada/fem/formats/sesam/read/reader.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import os
 
 from ada.api.spatial import Assembly, Part
+
 from .read_constraints import get_bcs, get_constraints
 from .read_elements import get_elements, get_mass, get_springs
 from .read_materials import get_materials
