+++ src/ada/visit/config.py
@@ -5,6 +5,9 @@ from typing import Callable, List, Union
 
 import numpy as np
 
+import ada.base.physical_objects
+from ada.visit.plots import build_display
+
 
 @dataclass
 class DataFilter:
@@ -34,3 +37,19 @@ class ExportConfig:
 
     metadata: dict = field(default_factory=dict)
     name_prefix: str = ""
+
+
+JUPYTER_GEOM_RENDERER: Callable[[ada.base.physical_objects.BackendGeom], None] | None = None
+JUPYTER_SECTION_RENDERER: Callable[[ada.sections.Section], None] | None = build_display
+
+
+def set_jupyter_part_renderer(renderer: Callable[[ada.base.physical_objects.BackendGeom], None]) -> None:
+    global JUPYTER_GEOM_RENDERER
+
+    JUPYTER_GEOM_RENDERER = renderer
+
+
+def set_jupyter_section_renderer(renderer: Callable[[ada.sections.Section], None]) -> None:
+    global JUPYTER_SECTION_RENDERER
+
+    JUPYTER_SECTION_RENDERER = renderer
