+++ tests/full/conftest.py
@@ -14,11 +14,11 @@ def dummy_display_func(ada_obj):
         logging.error("pythreejs is not installed. Install with 'conda install pythreejs'")
         return None
 
-    from ada.visit.rendering.renderer_pythreejs import MyRenderer, SectionRenderer
+    from ada.visit.plots import build_display
+    from ada.visit.rendering.renderer_pythreejs import MyRenderer
 
-    if type(ada_obj) is ada.Section:
-        sec_render = SectionRenderer()
-        _, _ = sec_render.build_display(ada_obj)
+    if isinstance(ada_obj, ada.Section):
+        build_display(ada_obj)
     else:
         renderer = MyRenderer()
         renderer.DisplayObj(ada_obj)
